fix: fix asset download 403 error
If a GitHub token is passed to a GitHub API of a public repo it can result in 403
if the org behind it has IP filtering enabled.

`dsaltares/fetch-gh-release-asset` does not support not passing GH tokens, so this
PR will switch to a more barebones approach of calling cURL directly to fetch the release.